+++ tests/utils/storage.py
@@ -3,8 +3,8 @@ from urllib.parse import urlparse
 from typing_extensions import override
 
 from aidial_adapter_openai.dial_api.resource import ValidationError
+from aidial_adapter_openai.dial_api.storage import Bucket, FileStorage
 from aidial_adapter_openai.utils.auth import Auth
-from aidial_adapter_openai.utils.storage import Bucket, FileStorage
 
 
 class MockFileStorage(FileStorage):
